split iota view header into ints view header and iota view header
