+++ include/range/v3/view/iota.hpp
@@ -424,83 +424,6 @@ namespace ranges
             {
                 constexpr auto&& closed_iota = static_const<closed_iota_fn>::value;
             }
-
-            struct ints_fn
-              : iota_view<int>
-            {
-                ints_fn() = default;
-
-                template<typename Val,
-                    CONCEPT_REQUIRES_(Integral<Val>())>
-                iota_view<Val> operator()(Val value) const
-                {
-                    return iota_view<Val>{value};
-                }
-                template<typename Val>
-                meta::if_c<
-                    (bool)Integral<Val>(),
-                    detail::take_exactly_view_<iota_view<Val>, true>>
-                operator()(Val from, Val to) const;
-
-            #ifndef RANGES_DOXYGEN_INVOKED
-                template<typename Val,
-                    CONCEPT_REQUIRES_(!Integral<Val>())>
-                void operator()(Val) const
-                {
-                    CONCEPT_ASSERT_MSG(Integral<Val>(),
-                        "The object passed to view::ints must be Integral");
-                }
-                template<typename Val,
-                    CONCEPT_REQUIRES_(!Integral<Val>())>
-                void operator()(Val, Val) const
-                {
-                    CONCEPT_ASSERT_MSG(Integral<Val>(),
-                        "The object passed to view::ints must be Integral");
-                }
-            #endif
-            };
-
-            template<typename Val>
-            meta::if_c<
-                (bool)Integral<Val>(),
-                detail::take_exactly_view_<iota_view<Val>, true>>
-            ints_fn::operator()(Val from, Val to) const
-            {
-                return {iota_view<Val>{from}, detail::iota_minus(to, from)};
-            }
-
-            struct closed_ints_fn
-            {
-                template<typename Val,
-                    CONCEPT_REQUIRES_(Integral<Val>())>
-                detail::take_exactly_view_<iota_view<Val>, true> operator()(Val from, Val to) const
-                {
-                    return {iota_view<Val>{from}, detail::iota_minus(to, from) + 1};
-                }
-            #ifndef RANGES_DOXYGEN_INVOKED
-                template<typename Val,
-                    CONCEPT_REQUIRES_(!Integral<Val>())>
-                void operator()(Val, Val) const
-                {
-                    CONCEPT_ASSERT_MSG(Integral<Val>(),
-                        "The object passed to view::closed_ints must be Integral");
-                }
-            #endif
-            };
-
-            /// \relates ints_fn
-            /// \ingroup group-views
-            namespace
-            {
-                constexpr auto&& ints = static_const<ints_fn>::value;
-            }
-
-            /// \relates closed_ints_fn
-            /// \ingroup group-views
-            namespace
-            {
-                constexpr auto&& closed_ints = static_const<closed_ints_fn>::value;
-            }
         }
         /// @}
     }
