+++ test/action/insert.cpp
@@ -10,6 +10,7 @@
 #include <set>
 #include <vector>
 #include <range/v3/core.hpp>
+#include <range/v3/view/ints.hpp>
 #include <range/v3/view/iota.hpp>
 #include <range/v3/view/take.hpp>
 #include <range/v3/view/for_each.hpp>
