+++ test/action/remove_if.cpp
@@ -9,7 +9,7 @@
 
 #include <vector>
 #include <range/v3/core.hpp>
-#include <range/v3/view/iota.hpp>
+#include <range/v3/view/ints.hpp>
 #include <range/v3/algorithm/move.hpp>
 #include <range/v3/action/sort.hpp>
 #include <range/v3/action/remove_if.hpp>
