+++ test/action/sort.cpp
@@ -11,7 +11,7 @@
 #include <random>
 #include <vector>
 #include <range/v3/core.hpp>
-#include <range/v3/view/iota.hpp>
+#include <range/v3/view/ints.hpp>
 #include <range/v3/view/stride.hpp>
 #include <range/v3/view/take.hpp>
 #include <range/v3/algorithm/shuffle.hpp>
