+++ test/action/unique.cpp
@@ -10,7 +10,7 @@
 #include <random>
 #include <vector>
 #include <range/v3/core.hpp>
-#include <range/v3/view/iota.hpp>
+#include <range/v3/view/ints.hpp>
 #include <range/v3/view/repeat_n.hpp>
 #include <range/v3/view/for_each.hpp>
 #include <range/v3/view/take.hpp>
