+++ test/algorithm/equal_range.cpp
@@ -24,8 +24,8 @@
 #include <range/v3/core.hpp>
 #include <range/v3/view/join.hpp>
 #include <range/v3/view/transform.hpp>
-#include <range/v3/view/iota.hpp>
 #include <range/v3/view/repeat_n.hpp>
+#include <range/v3/view/ints.hpp>
 #include <range/v3/view/take.hpp>
 #include <range/v3/utility/iterator.hpp>
 #include <range/v3/algorithm/copy.hpp>
