+++ test/distance.cpp
@@ -15,7 +15,7 @@
 #include <vector>
 #include <limits>
 #include <range/v3/core.hpp>
-#include <range/v3/view/iota.hpp>
+#include <range/v3/view/ints.hpp>
 #include <range/v3/view/take_while.hpp>
 #include "./simple_test.hpp"
 #include "./test_utils.hpp"
