+++ test/to_container.cpp
@@ -13,6 +13,7 @@
 #include <vector>
 #include <range/v3/core.hpp>
 #include <range/v3/to_container.hpp>
+#include <range/v3/view/ints.hpp>
 #include <range/v3/view/iota.hpp>
 #include <range/v3/view/transform.hpp>
 #include <range/v3/view/take.hpp>
