+++ test/view/any_view.cpp
@@ -10,7 +10,7 @@
 // Project home: https://github.com/ericniebler/range-v3
 
 #include <range/v3/core.hpp>
-#include <range/v3/view/iota.hpp>
+#include <range/v3/view/ints.hpp>
 #include <range/v3/view/take.hpp>
 #include <range/v3/view/reverse.hpp>
 #include <range/v3/view/any_view.hpp>
