+++ test/view/cycle.cpp
@@ -18,7 +18,6 @@
 #include <range/v3/algorithm/count_if.hpp>
 #include <range/v3/view/cycle.hpp>
 #include <range/v3/view/take.hpp>
-#include <range/v3/view/iota.hpp>
 #include <range/v3/view/reverse.hpp>
 #include <range/v3/view/c_str.hpp>
 #include "../simple_test.hpp"
